to tpl that speech recognition optimized
