+++ app/src/main/java/com/dvsnier/wrapper/TransferStationWrapper.java
@@ -28,7 +28,7 @@ import com.dvsnier.test.widget.theme.TestThemeActivity;
 import com.dvsnier.testEventBus.TestEventBusActivity;
 import com.dvsnier.test.tpl.green.TestGreenDaoActivity;
 import com.dvsnier.testOkhttp.TestOkhttpActivity;
-import com.dvsnier.testSpeechRecognition.TestSpeechRecognitionActivity;
+import com.dvsnier.test.tpl.speech.TestSpeechRecognitionActivity;
 import com.orhanobut.logger.Logger;
 
 /**
