+++ research/tpl/src/main/AndroidManifest.xml
@@ -17,5 +17,9 @@
             android:name=".green.TestGreenDaoActivity"
             android:theme="@style/AppTheme.DefaultTheme"
             android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
+        <activity
+            android:name=".speech.TestSpeechRecognitionActivity"
+            android:theme="@style/AppTheme.DefaultTheme"
+            android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
     </application>
 </manifest>
