+++ research/tpl/src/main/assets/iflytek/warning.png
