+++ research/tpl/src/main/res/layout/activity_test_speech_recognition.xml
