+++ research/widget/src/main/res/layout/activity_test_coordinator_layout_3.xml
@@ -21,7 +21,7 @@
             <ImageView
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:background="@drawable/tool_bar"
+                android:background="@android:drawable/list_selector_background"
                 android:scaleType="fitXY"
                 tools:ignore="ContentDescription" />
